% Auto-generated: do not edit by hand
\docType{package}
\name{dashCanvas-package}
\alias{dashCanvas}
\title{Interactive Image Annotation and Processing Component for Dash
}
\description{
Dash Canvas is a module for image annotation and image processing using Dash. It provides both the DashCanvas object for drawing and annotations on images, and a set of utility functions to process images using the annotations.    

}
\author{
\strong{Maintainer}: Emmanuelle Gouillart <emmanuelle.gouillart@nsup.org>
}
