% Auto-generated: do not edit by hand
\name{dashCanvas}

\alias{dashCanvas}

\title{DashCanvas component}

\description{
Canvas component for drawing on a background image and selecting regions.
}

\usage{
dashCanvas(id=NULL, image_content=NULL, zoom=NULL, width=NULL, height=NULL, scale=NULL, tool=NULL, lineWidth=NULL, lineColor=NULL, goButtonTitle=NULL, filename=NULL, trigger=NULL, json_data=NULL, hide_buttons=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks}

\item{image_content}{Image data string, formatted as png or jpg data string. Can be
generated by utils.io_utils.array_to_data_string.}

\item{zoom}{Zoom factor}

\item{width}{Width of the canvas}

\item{height}{Height of the canvas}

\item{scale}{Scaling ratio between canvas width and image width}

\item{tool}{Selection of drawing tool, among ["pencil", "pan", "circle",
"rectangle", "select", "line"].}

\item{lineWidth}{Width of drawing line (in pencil mode)}

\item{lineColor}{Color of drawing line (in pencil mode). Can be a text string,
like 'yellow', 'red', or a color triplet like 'rgb(255, 0, 0)'.
Alpha is possible with 'rgba(255, 0, 0, 0.5)'.}

\item{goButtonTitle}{Title of button}

\item{filename}{Name of image file to load (URL string)}

\item{trigger}{Counter of how many times the save button was pressed
(to be used mostly as input)}

\item{json_data}{Sketch content as JSON string, containing background image and
annotations. Use utils.parse_json.parse_jsonstring to parse
this string.}

\item{hide_buttons}{Names of buttons to hide. Names are "zoom", "pan", "line", "pencil",
"rectangle", "undo", "select".}
}
